Time: 6 ms (85.06%) | Memory: 46.5 MB (54.89%) - LeetSync

+++ 803-cheapest-flights-within-k-stops/cheapest-flights-within-k-stops.java
@@ -45,7 +45,7 @@ class Solution {
                 }
             }
         }
-        if(dist[dst]==1e9) return -1;
+        if(dist[dst]==(int)(1e9)) return -1;
         return dist[dst];
     }
 }
